# Make computing snapshot checksums faster in debug