--- a/lib/collection/src/tests/sha_256_test.rs
+++ b/lib/collection/src/tests/sha_256_test.rs
@@ -11,7 +11,7 @@ async fn test_sha_256_digest() -> std::io::Result<()> {
     let result_hash = hash_file(file.path()).await?;
     assert_eq!(
         result_hash,
-        "735e3ec1b05d901d07e84b1504518442aba2395fe3f945a1c962e81a8e152b2d"
+        "735e3ec1b05d901d07e84b1504518442aba2395fe3f945a1c962e81a8e152b2d",
     );
     Ok(())
 }