Computing file checksums is extremely slow in our debug builds. I was experimenting with snapshots again today and decided to fix this because it was quite annoying.

It now uses `sha256sum` which is significantly faster. I only enabled this for debug builds because: 1. in release hashing is fast already. 2. you better don't call random binaries in release.

Before: checksum takes 20s, snapshot takes 40s
After: checksum takes 0.5s, snapshot takes 20s

Details:
- only used in debug builds, on Unix platforms
- may fail if `sha256sum` doesn't exist somehow, falls back to built-in hasher
- includes a test to ensure both hashers give consistent results

### All Submissions:

* [x] Contributions should target the `dev` branch. Did you create your branch from `dev`?
* [x] Have you followed the guidelines in our Contributing document?
* [x] Have you checked to ensure there aren't other open [Pull Requests](../../../pulls) for the same update/change?

<!-- You can erase any parts of this template not applicable to your Pull Request. -->

### New Feature Submissions:

1. [x] Does your submission pass tests?
2. [x] Have you formatted your code locally using `cargo +nightly fmt --all` command prior to submission?
3. [x] Have you checked your code using `cargo clippy --all --all-features` command?