@ffuugoo 2024-09-11
On macOS it's just `shasum`... 😂
<img width="554" alt="Screenshot 2024-09-11 at 11 39 40" src="https://github.com/user-attachments/assets/c1efe618-d5b3-4aed-8681-37a4ada09971">

Though, I don't think we care about macOS much. 