@timvisee 2024-09-11
> On macOS it's just `shasum`... 😂 <img alt="Screenshot 2024-09-11 at 11 39 40" width="554" src="https://private-user-images.githubusercontent.com/2725918/366395922-c1efe618-d5b3-4aed-8681-37a4ada09971.png?jwt=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3MjYwNDgyMDEsIm5iZiI6MTcyNjA0NzkwMSwicGF0aCI6Ii8yNzI1OTE4LzM2NjM5NTkyMi1jMWVmZTYxOC1kNWIzLTRhZWQtODY4MS0zN2E0YWRhMDk5NzEucG5nP1gtQW16LUFsZ29yaXRobT1BV1M0LUhNQUMtU0hBMjU2JlgtQW16LUNyZWRlbnRpYWw9QUtJQVZDT0RZTFNBNTNQUUs0WkElMkYyMDI0MDkxMSUyRnVzLWVhc3QtMSUyRnMzJTJGYXdzNF9yZXF1ZXN0JlgtQW16LURhdGU9MjAyNDA5MTFUMDk0NTAxWiZYLUFtei1FeHBpcmVzPTMwMCZYLUFtei1TaWduYXR1cmU9ZjIxZGNlZmEzZmYxZDk2OTUyMTFkMDJmYTZkMWVmNGI0M2VlZjdkOTAzZWUwMjk0YmE5MDg2MzM1M2U4YzBmYyZYLUFtei1TaWduZWRIZWFkZXJzPWhvc3QmYWN0b3JfaWQ9MCZrZXlfaWQ9MCZyZXBvX2lkPTAifQ.SmkpZw4gvWLGY0ZceXnmxk2CqbkH5HwsKcRjQ8nIf-8">
> 
> Though, I don't think we care about macOS much.

Oh noes!

Thanks for checking. I'll validate on my end how likely it is for `sha256sum` to exist. Changing the compile time conditions is probably good enough.