@timvisee 2024-09-11
I asked in our internal channels, and all Linux users do seem to have it.

Linux does also have `shasum` but it's about 2.5 times slower in my quick tests.

I'll just enable it on Linux only then to keep things simple. It's not very significant anyway and falls back if the binary is not there.