@ffuugoo 2024-09-12
> We could go with [`Cargo.toml` overrides](https://doc.rust-lang.org/cargo/reference/profiles.html#overrides), i.e. by adding this to our top-level `Cargo.toml`:
> 
> ```toml
> [profile.dev.package.sha2]
> opt-level = 3
> ```

🤯

I vote to test this one first, then.