@timvisee 2024-09-12
> > Ideally we enable release optimizations for the built-in hasher at all times. But I don't see a proper way to achieve this.
> 
> We could go with [`Cargo.toml` overrides](https://doc.rust-lang.org/cargo/reference/profiles.html#overrides), i.e. by adding this to our top-level `Cargo.toml`:
> 
> ```toml
> [profile.dev.package.sha2]
> opt-level = 3
> ```

Woah! I wasn't aware of this 🥹

Let me give it a quick shot. I'll definitely go with that if that works.