@timvisee 2024-09-12
Brilliant, works like a charm!

Latest test for hashing duration:
- before: 15.6s
- with `sha256sum`: 0.3s (got to love GNU coretutils)
- with release mode: 0.4s

That's obviously good enough. I've updated the PR to enable release optimizations for the sub-crate.